Move admin commands to seperate folder
