+++ src/MomentumDiscordBot/MomentumDiscordBot/Discord/Commands/Admin/AdminConfigModule.cs
@@ -1,14 +1,12 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.ComponentModel;
 using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 using Discord;
 using Discord.Commands;
 using MomentumDiscordBot.Models;
 
-namespace MomentumDiscordBot.Discord.Commands
+namespace MomentumDiscordBot.Discord.Commands.Admin
 {
     [Group("config")]
     public class AdminConfigModule : AdminModule
