+++ src/MomentumDiscordBot/MomentumDiscordBot/Discord/Commands/Admin/AdminTwitchBanModule.cs
@@ -4,10 +4,9 @@ using System.Threading.Tasks;
 using Discord;
 using Discord.Commands;
 using MomentumDiscordBot.Models;
-using MomentumDiscordBot.Services;
 using MomentumDiscordBot.Utilities;
 
-namespace MomentumDiscordBot.Discord.Commands
+namespace MomentumDiscordBot.Discord.Commands.Admin
 {
     [Group("twitchBan")]
     public class AdminTwitchBanModule : AdminModule
